requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5
drf-nested-routers>=0.93.4

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Authentication
djangorestframework-simplejwt>=5.3.0
django-allauth>=0.59.0
dj-rest-auth>=5.0.2

# Media Storage
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0

# Security
python-decouple>=3.8

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0

# Utilities
Pillow>=10.1.0
python-dateutil>=2.8.2

# Celery (for async tasks)
celery>=5.3.4
redis>=5.0.1
django-celery-beat>=2.5.0
django-celery-results>=2.5.1

# Development
django-extensions>=3.2.3
